requirements.txt
argcomplete==1.10.3
azure-ai-formrecognizer==3.3.3
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.30.1
beautifulsoup4==4.8.2
blinker==1.8.2
bs4==0.0.2
certifi==2024.6.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
compoundfiles==0.3
compressed_rtf==1.0.6
docx2txt==0.8
easyocr==1.7.1
ebcdic==1.1.1
eml-parser==1.17.5
et-xmlfile==1.1.0
extract-msg==0.28.7
filelock==3.14.0
Flask==3.0.3
Flask-Cors==4.0.1
fsspec==2024.6.0
idna==3.7
imageio==2.34.1
IMAPClient==2.1.0
intel-openmp==2021.4.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
lazy_loader==0.4
lxml==5.2.2
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
msg-parser==1.2.0
msrest==0.7.1
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
opencv-python-headless==4.10.0.82
openpyxl==3.1.3
outlook-msg==1.0.0
packaging==24.0
pandas==2.2.2
pdfminer.six==20191110
pillow==10.3.0
plum-dispatch==1.7.4
publicsuffixlist==0.10.1.20240605
pyclipper==1.3.0.post5
pycryptodome==3.20.0
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pypandoc_binary==1.13
pypdf==4.2.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
python-pptx==0.6.23
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
scikit-image==0.23.2
scipy==1.13.1
shapely==2.0.4
six==1.12.0
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.8.1
Spire.Doc==12.4.0
sympy==1.12.1
tbb==2021.12.0
tifffile==2024.5.22
torch==2.3.1
torchvision==0.18.1
typing_extensions==4.12.1
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
Werkzeug==3.0.3
xlrd==1.2.0
XlsxWriter==3.2.0
